Implemented weapon swap
Weapon swapping is now (mostly) supported. I don't have a Blunt<->Edged
weapon to test with, but it works for Twohanded<->NotTwoHanded cases.

Changed as part of this was removing support for the case in which
you're already wielding the weapon when the next skill is chosen.

Changed as part of this is increasing the target skill value from
"current + 2" to "current + 3".