Register Bow as a two-handed skill
This fixes the case in which a bow is NOT a worn weapon.

When a bow is NOT worn, a wield command is used. The result of the wield
command is that of a two-handed weapon, which the swap logic detected
and swapped the bow into the left hand.

I suspect Crossbow needs the same fix, but I do not have one with which
to test.