Handle failed charged maneuver use
Extra pause and roundtime checks are needed when a charged maneuver is
attempted and succeeds. Now the charged maneuver logic detects a failure
and does not wait.

The cooldown timer logic could be improved. Right now when a failure is
detected it resets the cooldown timer. We should probably instead do
something like add 10 seconds to it.

Fixes #41.