[WIP] Add alternate attack logic.
Can I pull the variables in the same way as skinning does, or do I have to go through the equipment manager?  Now that I'm thinking about it, I probably need to pull the attack_type in when each weapon changes.

Can one of you point me to the weapon change method?