Fighting the good fight against random drops.
[Talas Sagamat, Northern Tunnel - 8296]
Propped against one side of the cave is a mostly intact suit of rusted
plate armor, the skeletal remains of its owner still holding a broken
sword in one of its boney hands.  Tangled between the fingers of the
other is a mass of brownish fur.  A small phrase has been deeply
scratched into the armor's surface, in memory of those who perished
along with the warrior.  You also see a quartz gargoyle, a quartz
gargoyle that is lying down, a twisted anlora-avtoma armband and some
junk.
Obvious exits: north, south.
>
[combat-trainer]>invoke armband
You should probably be holding a twisted anlora-avtoma armband to do
that.