modified stow_lootables function in combat-trainer
Modified stow_lootables function in combat-trainer to attempt to stow loot items in the picking_box_storage container set in YAML before declaring it unlootable and dropping it.  If picking_box_storage is not set or unable to store in set container, it finishes declaring it unlootable and dropping it.

I set this up mainly for boxes, as pick.lic will look in pick_box_storage after picking_box_source is empty. I don't see it unsustainably filling the container with loot. However, perhaps a function somewhere to clean up this container might be necessary. IE a fill my gempouch with my pick_box_storage?