modified stow_lootables function in combat-trainer
Modified stow_lootables function in combat-trainer to attempt to stow loot items in the picking_box_storage container set in YAML before declaring it unlootable and dropping it. If picking_box_storage is not set or unable to store in set container, it finishes declaring it unlootable and dropping it. 

This only goes when store_extra_loot is set to true in YAML