support minimum mana levels in hunting rooms
Altered common-travel find_empty_room so that you can define a minimum
mana level as well as how strictly you care about that. If it finds an
empty room, it will check the mana level. If it does not hit the
threshold, it moves on.  In the event that it finds empty rooms, but
they do not have the appropriate mana level, another argument will
determine if you should idle waiting for a good mana empty room (strict
mode) or just fall back on any empty room (non strict).