Add more release cyclics (#1523)
* add release_cyclics to before_dying

Lots of WMs have found themselves killing hunting-buddy and/or combat-trainer and then walking to town with EE running and get jutsice violations. This fixes that.

* add before_dying, and put release_cyclics

Lots of WMs have found themselves killing hunting-buddy and/or combat-trainer and then walking to town with EE running and get jutsice violations. This fixes that.