Handles pouring holy oil during forging
This change allows Clerics (and others) who pour holy oil, rather than
the flask of oil, during forging to continue searching for the proper
oil to finalize their product. The stow_item('oil') in the while loop
will not stow anything if the holy oil is all poured out. But the script
should continue and the item should be finalized.