Fix holy oil / alchemist oil error in forge (#1558)
* Fix holy oil errors in forging

The pour_oil method routinely picked up holy oil, poured it instead of
the oil necessary to finish a forged item, and ended up unable to finish
the item. I tested the attached change with multiple instances of holy
oil and normal oil from the alchemist. It handled the issue (though it
may pour out all of your holy oil/oil).

I have not tested it against an oil that gives the message "You can't
pour". I do not know where to find such an oil. But it should follow the
same procedure as the original pour_oil method when it encounters this
issue.

* Add Protection from Evil to base-spells