Update to Steal.lic to support other hometowns
Selects stealing targets based on user's hometown selection.
Changed the build_new function variable province to steal_town.  The variable was static before and did not change.  I did not find any place where the province variable was being utilized.  Due to this change in granularity and variable name change, base-stealing.yaml would need to be updated with new values.  I used find and replace to change all of the "province: zoluren" to "steal_town: Crossing" in the base-stealing.yaml.