Fixed 'pets' not tracking # of pets
In the attempt_open function 'if @make_pets' was getting called but did not match when the pet box was stored in the picking_pet_box_source container. Therefore @pet_count was never increased and it would turn all source boxes into pet boxes. The issue is fixed with this change.

This is the subject of issue ticket #1561.