Fixed 'pets' not tracking # of pets (#1573)
In the attempt_open function 'if @make_pets' was getting called but did not match when the pet box was stored in the picking_pet_box_source container. Therefore @pet_count was never increased (@pet_count += 1) and it would turn all source boxes into pet boxes. The issue is fixed with this change because 'You put your' is matching correctly.

Fixes #1561