I removed the entry:
name: a metal flask
  noun: flask
  volume: 5
  type: blacksmithing
  work_order: true
  chapter: 6

because "a metal flask" results in a bad match when called... the appropriate entry is "a small metal flask"... and when forge.lic called the inappropriately labeled item, it results in a failure; evidence:

forge]>turn my book to chapter 6
You turn your book to chapter 6, entitled "Forged Item Design".
>
[forge]>read my book
              -=    Chapter 6 : Forged Item Design    =-

 Page 1: a slender metal rod                     Page 9: a back scratcher
 Page 2: a shallow metal cup                     Page 10: a metal bolt box
 Page 3: a metal horseshoe                       Page 11: a metal origami case
 Page 4: a soft metal keyblank                   Page 12: a metal lockpick case
 Page 5: a short metal mug                       Page 13: a metal ankle band
 Page 6: a tall metal mug                        Page 14: a small metal flask
 Page 7: a large metal horseshoe
 Page 8: some metal barbells
>
Your circle of obsidian on your forehead cools slightly.
>
[forge: *** No match was found after 15 seconds, dumping info]
[forge: messages seen length: 10]
[forge: message:               -=    Chapter 6 : Forged Item Design    =-]
[forge: message:  Page 1: a slender metal rod                     Page 9: a back scratcher]
[forge: message:  Page 2: a shallow metal cup                     Page 10: a metal bolt box]
[forge: message:  Page 3: a metal horseshoe                       Page 11: a metal origami case]
[forge: message:  Page 4: a soft metal keyblank                   Page 12: a metal lockpick case]
[forge: message:  Page 5: a short metal mug                       Page 13: a metal ankle band]
[forge: message:  Page 6: a tall metal mug                        Page 14: a small metal flask]
[forge: message:  Page 7: a large metal horseshoe                 ]
[forge: message:  Page 8: some metal barbells                     ]
[forge: message: Your circle of obsidian on your forehead cools slightly.]
[forge: checked against [/Page \d+.*a metal flask/i]]
[forge]>turn my book to page
Which page do you wish to turn the a book of apprentice blacksmithing instructions to?
>
[forge: *** No match was found after 15 seconds, dumping info]
[forge: messages seen length: 1]
[forge: message: Which page do you wish to turn the a book of apprentice blacksmithing instructions to?]
[forge: checked against [/You turn your/i, /The book is already/i]]
[forge]>study my book
You quickly flip through the book and determine it is a collection of instructions for the blacksmithing crafting Discipline in the Forging crafting society.

Upon further perusal, it appears you can TURN the BOOK to the FOREWORD, and read a short primer on how crafting of this type is performed.  You may also TURN the BOOK to INDEX, and read it for a list of chapters.  Then, you can TURN the BOOK to CHAPTER # and read it for a list of instructions on each page.  Finally, you can TURN the BOOK to a PAGE # in that chapter, and read it for detailed information on what is required for the crafting process.

Once you have turned to the page you wish to craft, you may STUDY the BOOK to begin the crafting process.

With some complex books, it is possible to TURN the BOOK to [discipline name] to select a different set of templates.
>
[forge: *** No match was found after 15 seconds, dumping info]
[forge: messages seen length: 5]
[forge: message: You quickly flip through the book and determine it is a collection of instructions for the blacksmithing crafting Discipline in the Forging crafting society.]
[forge: message: Upon further perusal, it appears you can TURN the BOOK to the FOREWORD, and read a short primer on how crafting of this type is performed.  You may also TURN the BOOK to INDEX, and read it for a list of chapters.  Then, you can TURN the BOOK to CHAPTER # and read it for a list of instructions on each page.  Finally, you can TURN the BOOK to a PAGE # in that chapter, and read it for detailed information on what is required for the crafting process.]
[forge: message: Once you have turned to the page you wish to craft, you may STUDY the BOOK to begin the crafting process.]
[forge: message: With some complex books, it is possible to TURN the BOOK to [discipline name] to select a different set of templates.]
[forge: message: [DRPrime]-DR:Douken: "Is there an example Empath YAML available?"]
[forge: checked against [/Roundtime/i]]
[forge]>put my book in my bag
You put your book in your tool bag.
>
[forge]>get my bronze ingot
>;chat yup that is the issue runningto it again
You get a bronze ingot from inside your hunting pack.
>
[forge]>put my ingot on anvil
You put your ingot on the iron anvil.
>
[forge]>untie my forging hammer from my forger belt
You remove a polished forging hammer crafted from engraved icesteel from your forger's belt.
>
[forge]>untie my tongs from my forger belt
You remove some polished tongs crafted from engraved icesteel from your forger's belt.
>
[forge]>pound ingot on anvil with my hammer
You cannot figure out how to do that.  Perhaps finding suitable ingredients and studying some instructions would help.
>;pause
--- Lich: forge paused.