Enable bankbot when bank is disabled
Change allows use of bankbot when bank bypass is on. This will also deposit all coins with the bankbot since exchange is disabled with bank bypass. Allows threshhold function of local currency while depositing all others.

Usage is activated if bank bypass is on and a bankbot name is present.

if settings.sell_loot_skip_bank && settings.bankbot_name