Fix dancing (#1782)
Dancing currently skips your first action (and third action and subsequent odd numbered actions. The current version calls next_dance_action twice at line 1745 so the first action is removed in the "unless" statement and the second dance action immediately runs.

This change fixes the problem. The "unless" at line 1745 is not necessary because line 1744 already checks if the @dance_queue is empty and, if so, creates a new queue.