Communes (#1793)
* Add cast messaging for Sanyu Lyba

Add cast messaging for Cleric spell Sanyu Lyba (You open your arms in a
hieratic gesture of summoning.)

* Ensure water_holder placed in theurgy_container

In both CT and theurgy.lic the @water_holder is put into the
@theurgy_supply_container in all instances other than this single
scenario in CT.

* water_holder to/from theurgy_supply_container

In Meraud's commune the @water_holder should be gotten from, and put into, the @theurgy_supply_container.

* Handles pouring holy oil during forging

This change allows Clerics (and others) who pour holy oil, rather than
the flask of oil, during forging to continue searching for the proper
oil to finalize their product. The stow_item('oil') in the while loop
will not stow anything if the holy oil is all poured out. But the script
should continue and the item should be finalized.

* Revert "Handles pouring holy oil during forging"

This reverts commit a27a021274eca37d38f8f5e9ea6adff957d19ea9.

* Fix holy oil errors in forging

The pour_oil method routinely picked up holy oil, poured it instead of
the oil necessary to finish a forged item, and ended up unable to finish
the item. I tested the attached change with multiple instances of holy
oil and normal oil from the alchemist. It handled the issue (though it
may pour out all of your holy oil/oil).

I have not tested it against an oil that gives the message "You can't
pour". I do not know where to find such an oil. But it should follow the
same procedure as the original pour_oil method when it encounters this
issue.

* Add Protection from Evil to base-spells

* Fixed 'pets' not tracking # of pets

In the attempt_open function 'if @make_pets' was getting called but did not match when the pet box was stored in the picking_pet_box_source container. Therefore @pet_count was never increased and it would turn all source boxes into pet boxes. The issue is fixed with this change.

This is the subject of issue ticket #1561.

* add COZ cast message

add cast messaging for Curse of Zachriedek

* Perform communes

To get description: ;commune
(ex: ;commune meraud will check for material and perform commune)

The script can also carve prayer beads if given the full name of an immortal or name of animal to carve.
(ex: ;commune carve kertigen (will carve a raven bead))

* changes per comments

* incorporated requested changes from 8/27/17

Changed get_cleric_item, stow_cleric_item, and have_cleric_item so they do not return true/false. Get_cleric_item now exits if item does not exist.

empty_cleric_hands simplified because the parameter was never called. It now only runs empty_cleric_right_hand and empty_cleric_left_hand.

Removed string interpolation in check_devotion because the value was already a string.