links.lic: Update to yaml parser to http://yaml-online-parser.appspot.com/
This parser gives much more information on a block error which is the most common error.

For instance, instead of this that lich reports:
while parsing a block mapping
  in "<unicode string>", line 3, column 1:
    hunting_info:

the parser will report:
while parsing a block mapping
  in "<unicode string>", line 3, column 1:
    hunting_info:
    ^
expected <block end>, but found '<block sequence start>'
  in "<unicode string>", line 7, column 2:
     - undead
     ^

where the second hunk is the actual problem that needs fixing.