smash-pumpkins.lic: Various updates.
Use the fix_standing helper from common.
Check for not enough coins to continue and withdraw from the bank, or exit if not enough coin.
switch an fput to a bput to put the item in the bucket
change the safe room to the lobby now that the entrance is magic free in 2017.