Add empty_hands to crossing-repair setup (#2063)
Based on certain factors (blindness, RT, etc) armor was in hand when calling crossing-repair. If that armor is not part of the standard gear set, it breaks crossing-repair when trying to retrieve the item for repair. Easy fix is by calling empty_hands after the standard gear set is equipped. This will make sure any other armor items are appropriately stowed.