fix exiting Junk Room
It appears there was a recent change where dropping 1 item doesn't always let you exit the Junk Room. This caused the ;steal script to stall in the Junk Room when dropping 1 item didn't allow an exit.

The proposed change looks at the portal after each dropped item until the portal indicates it is safe to exit.