Update Healme to use waggle sets, more spells, and current check_health method (#2139)
* Add flush poison, cure disease, and health check

The health check method is a predicate that returns true if poisoned,
diseased, or wounded. Flush poison and cure disease methods pull data
from a waggle set called 'healing'. They cast the appropriate spell
until disease/poison free.

* Guard the heal self method against non-empaths

There is no support for other guilds at the moment, so this guards
against non-empaths using it.

* Add support for waggle set 'healing'

This opens up the script to use the more standardized waggle_set spell
data formatting. The waggle_set is named 'healing' and will be the place
to list spells like cure disease, flush poisons, regenerate, and
fountains of creation.

* Support regenerate and casting heal via waggle_set

Check if regenerate is active before casting to save time. If the user
has 'heal' in waggle_sets['healing'] then we defer to that over the
empath_healing set.

* Call the flush poisons and cure disease methods

Call the disease/poison healing methods after casting regenerate and/or
heal. The order of this could be optimized in a future request.

* Use check_perc_health and return if unhurt

Removed the predicate method `need_healing?` in favor of
check_perc_health.

* Use check_health before poison/disease casting