base-crafting.yaml: Switch all shaping to use maple instead of balsa. (#2157)
Workorders pay better for maple, and shaping is not hardcoded to use balsa because it is 10 length (aka the max length any of the previous recipes would have needed) anymore now that it supports buying enough of any wood to complete any recipe. 80 workability for maple vs 85 for balsa is insignificant.