hunting-buddy.lic: Add during: option. (#2187)
This is to run arbitrary scripts only while in combat. For instance, I have one that summons a wolf companion, and watches for triggers to react to so you can automate the process of getting them older and dismisses it when the script is killed. Keeping the wolf out always is not an option. Baby wolves can't deliver items to other people, and it takes around 80 hours for them to get older so this seems like the most fitting place to add it. I'm sure it could be used for other things too.

Usage:
hunting_info:
- :zone: black_goblins
  args:
  - d1
  stop_on:
  - Small Edged
  - Stealth
  :duration: 50
  before:
  - mech-lore
  - athletics
  during:
  - combat-ranger-companion
  after:
  - athletics