Updated text2num to fix bug in parsing >=100 (#2198)
Because 'hundred' was not in the map (on purpose, for scaling) the nil condition was being met first. Switched it to be checked first. Limited testing works now. 

Before:
You count some small rocks in the quiver and see there are one hundred three left.                                                                                                                       
[restock: Unknown number]   

After:
You count some small rocks in the quiver and see there are one hundred three left.                                                                                                                       
[restock: 103]