Updating missing remedies item in base-recipies
Adding missing recipe "some neck tonic". It is only one of two difficulty 7 - challenging recipes where all ingredients are sold by the society. From the order of the remedies recipes, it looks like everything was just added in order from the Wiki, but this one was just skipped.

This had originally blocked having a viable workorder at any difficulty at 601 alchemy with no techniques. Stopgap is to switch to "easy" workorders for a time with this being the only option available.