Combat-Trainer: Add setting for allowing spell mana/cambrinth/harness data to be calculated dynamically (#2320)
* Add auto mana spell setting

This works the same as crossing-trainers calculation of mana when
training magic. Setting 'use_auto_mana' will use the the discerned data
for prepping the spell. Setting 'use_auto_mana' and 'symbiosis' will
search for the mana value which teaches according to the
symbiosis_learning_threshold setting.

* Fix a typo

* Store settings needed for using check_discern

* Handle multi-cambrinth data

* Modify charge setting

The cambrinth charge setting checks the first element of the data,
grabbing the first element if it is an array. This is to handle discern
data which returns an array of arrays representing the charge data.

* Require ostruct

Common-arcana::check_discern takes a bulk of user settings which is an
openstruct.

* Rubocop

* Store settings in instance var

* dont use get_settings