Add mm.lic, A general moon mage utility script (#2435)
* Add mm.lic

The initial commit contains the locate function, which locates a player
and displays the room number if the locate was successful.

* add doc link and remove common

* require events

* Add rift; Make locate a predicate

Adds rift functionality. Changed locate to a predicate so it could be
used by the rift method to decide whether or not to cast rift.

* Return locate retry value

locate? would return false otherwise

* Use correct method name

* Print error message instead of retry locate

There is potential for infinite looping if the caster cannot cast CV for
whatever reason.

* use methods for handling flags