shape.lic: Handle applying stain with shape continue
You deeply analyze the Nisha shortbow and pore over its construction.
When crafting is complete this will be a type of finished shortbow.
With all major carving work completed, the wood is ready for the application of stain.
From the progress so far, it looks like the shortbow is masterfully-crafted (12/12).
The shortbow is a somewhat challenging piece to make.
You estimate that it is very nearly done.
Roundtime: 8 sec.