Add validation for combat_spell_training
This got me pretty good. I was following the pattern set with `buff_spells` in naming the key as a spell name, so for a few days none of my magics were training in combat and I couldn't figure out why. Finally did, and decided to add in a validation to clue in others.