New changes
Ok, new changes:
;rezz by itself will raise all corpses in the room
;rezz <target> will only raise that person.

Put both Rejuv and Rezz in 1 waggle set, and now it gives a message if you don't have it setup.

Now the script only does the 'look people' if you run it without any arguments so it can check for dead people.

Made the methods more efficient and removed some of the unnecessary ones.

They're now called with rejuv_process and rezz_process.

If soul isn't found in 10 infuses, it gives the option to give up or keep going.  If you keep going, it will do 10 more and ask again as many times as you let it continue.

I added the script runtime displaying at the end because I was curious.  That can stay or go, whatever is appropriate.

The only thing I can think of that isn't handled is if the body is either dragged off or someone else raises them first.
If they get raised while you're still rejuving, it will deal with it.  But not after that.

I could add some more messages to the 'alive' flag.   "You can't cast that at yourself." will show up both if you try to rejuv or soul bond someone who isn't there.

Oh, just added one more flag for if the person is a necro.  I ran across a dead one while testing and noticed that they have a different message when you perceive them.