More corrections
I took care of all of the indentation issues, I hope.

I also fixed a few of the style issues as well as made the ternary if statements with begin/end into multi line if statements so that they're more readable.

I changed one of the flags so that the script doesn't have to attempt to rejuv before realizing that the player is already alive.
That change might require some code to handle if someone rezzes them while you're trying to find their soul.  Not certain.

Let me know if there are any other things I can improve!