Fix merging of settings and UserVars for safe_room_empaths
Because `safe_room_empaths` has an entry in `base-empty`, it would never reach this point and be `nil`.

Therefore the UserVar value would never be merged with the setting value.

The other transforms here are not merging anything with an entry in `base-empty`, so this has not been a problem before.