Be extra careful about starting/stopping performance.
On slow systems 2 seconds is not enough to catch the dirty message in performance for first-aid so increase that to 4.
Also add a 2 second pause if performance is still running, performance handles stop playing in before_dying, which happens after the script is stopped but before it's actually in exited status, causing problems when handing off to another script that runs performance.