rezz.lic: Delete rotting flag, update rejuv method (#2989)
Added a line to delete the rotting flag before dying (it's created in def find_soul?).

Edited the rejuv method.  It was overwriting the global @rejuv variable if it was called with quick = true.  This would only cause an issue if you're trying to rezz more than 1 body, ie: you called it with just ;rezz, and there are multiple bodies in the room.  It would be rejuving the second body with the quick settings.

Now it creates a variable in the method and sets it either to @rejuv or the quick rejuv settings.