Update ;pick to allow storing harvested trap components
If component_container: is defined in the yaml, then when a picker is harvesting components it will attempt to place them inside the container.  If it can't, it will drop/dispose of them as normal.

Also a bugfix that prevents a rare problem I've run into.  There is a situation where stowing pliable equipment (eg. removing and stowing all the hindering equipment before beginning picking) can fill picking_box_source to the point that if it runs into a box that is too hard to disarm it will fail to stow it.  At which point pick will continue trying to pull boxes.
  - Thief removes box from backpack
  - Thief fails to disarm box.
  - Thief tries to stow box but can't fit it inside the backpack anymore.
  - Thief grabs the next box. and has two in his/her hands.

In rarer situations this can cause a picker to try and pick an un-disarmed box.  

With this change if it fails to stow the box it will drop it and continue picking.