focus retreat behavior and correct mana_threshold
Reverting behavior w/regards to retreating except for initial hunt start when RT is involved with swapping real gear out for regalia.

Also, correcting edge case by adding buff_spell_mana_threshold check.  Although it's not exactly a buff and doesn't show up as an active spell, It was possible for regalia to try and cast without the mana for it, which would lead to a time without armor cover.