add razortusk boar hunt and organize ratha base-hunting
Added a set of spawn rooms for Ratha's razortusk_boars. 

Changed the elanthipedia link for the Kra'hei hatchlings -- the fully grown Kra'hei aren't on the list and will be added once I can safely explore in there.

Added all the mob hunt entries accessible from Ratha to the hunting_areas_by_town section.  Chose to include shadow_masters -- technically on Taisgath, but since you can walk there in moments and Taisgath's not even really a hometown it went with the rest of Ratha's stuff