fix ratha's sandsprite hunting zone
Ratha's sandsprite hunting zone has many entries that are offset by 300 from their real room ids.  This means that if the first room isn't free it will take you all the way to the Aesry graveyard.  I've hunted there for a long time and never noticed this, but ALL of the 55xx rooms are in Aesry, and improperly tagged with 'sandsprites' in their room data (I'll fix this in the map later)  There are no sandsprites outside of Ratha's wreck.

The real rooms are the -300 offset versions.  5521 becomes 5221, etc.  All of the rooms have been checked to make sure they are the right versions, so it looks like fumbled numbers when they were put in originally.