update bloodvines_hib hunting zone (#3134)
Added peer tags to map all of the identical rooms in bloodvines_hib hunting zone.  The mapping and waytos were redone and each room can be go2ed and LRs as the right Room ID.

Room 13058 is a relic of the past and now lfrs as being in Akigwe's Tower.  No other rooms are removed from the list.