Update base-hunting.yaml
I have added a hunting area for blood dryads.  I did not put all the rooms in as there are about 18-20 of them.  I used 7 as this does not seem to be a crowded area.  Also, note that Blood Nyads are in this area as well and this is noted.