;shape stamping and bow enhancement (#3168)
Related to Issue https://github.com/rpherbig/dr-scripts/issues/3166

Added stamping support to shape.

Added a check to make sure you're holding either the noun you're working on or raw lumber.  If it's in the right hand it swaps it to the left so it can be worked on.

Finally, added laminating, lightening, and cable-backing for bows.  Tested with Barder's help.