fix problem with empty analyze array on mob death (#3170)
* fix problem with empty analyze array on mob death

This is related to issue https://github.com/rpherbig/dr-scripts/issues/3143

A condition can trigger whenever a mob is killed at the same time as a barbarian's analyze array is depleted and not replenished by perform_analyze? that causes an empty melee_attack_verb to be generated.

I don't believe there's an easy fix to this 'in-method' for perform_analyze?, since it would need to handle the states of being out of melee or all mobs being dead, and do it before a melee_attack_verb is decided.  So instead whereever update_analyze_array is used it will just default to 'attack' if it returns with an empty array.

Also, for somewhat more reliability, the face next in perform_analyze does try once more to turn on an enemy in melee, which should prevent this most of the time.

* fix mismatched match message

Due to idiocy, the when didn't match the case.