Undo abbrev spell-checking
After some time talking to people I think the name-checking portion of this is doing its job, but the number of people with incomplete or non-matching abbrevs in their yamls is pretty high, especially for the spell abbrevs not explicitly defined in SPELLS.  This seems to indicate to me it probably isn't a necessary check and may obscure real problems in validate when working with people.

I'd like to revert only the abbrev checking and keep the name checking as is.