update logic for exiting from junk yard
The old logic checked if a user was still in the junk yard 60 seconds after entering. The new logic only exits the game if you don't leave the junk yard for a *consecutive* 60 seconds.

People on the verge of item limits can go in and out of the junk yard frequently. This fixes the situation where a player exits because they happen to be back in the junk yard 60 seconds after initially triggering the ;afk script.