Addition of ;tinker script
Beginning support for Tinkering.

This script was made from ;carve and conforms to the same style.

Currently the args look like:

```
  ;tinker <finish> <chapter> <recipe name> <material> <noun> [mechanism]
   finish       What to do with the finished item. [log, stow, trash]
   chapter      Chapter containing the item.
   recipe name  Name of the recipe, wrap in double quotes if this is multiple words.
   material     Type of material.
   noun
   mechanism    What metal mechanism to use -brass-oravir-etc. Optional.

  ;tinker <continue> <noun>
   continue
   noun

  ;tinker laminate <noun>
   laminate
   noun         Noun of item to laminate.

  ;tinker lighten <noun>
   lighten
   noun         Noun of item to lighten.

  ;tinker cable <noun>
   cable
   noun         Noun of item to cable-back and increase draw-strength.

  ;tinker mechanisms <material>
   mechanisms   Walk to an open gear press and make a single stack of mechanisms from an existing ingot.
   material     What type of metal ingot to use
```

It has been tested with all types of crossbows/stonebows as well as tinkered toys and mechanisms. Also added were a tinkering_tools: entry to base and base-empty, as well as an entry in the no_use scripts for sano and almanac stuff.

It does not currently buy parts nor are the recipes in base-recipes.