Update afk.lic - Get all items at feet!! (#3277)
* Update afk.lic - Get all items at feet!!

Added some code to check if it's 1 item or multiple items at your feet.  If multiple, it builds a list of the items and gets them all.

I also changed it to pause go2 and then unpause it at the end.  I'm not sure if this is okay with the overall intent here.
I don't know if there's a way to send a signal to go2 to tell it to restart without waiting 30 seconds.
If there is, please let me know.

Or if it's preferred that it just stops go2, I'll change that part back.

* Update afk.lic

* Update afk.lic

I removed the lines to pause and unpause go2 and put the one to stop it back in.