Support room declaration by town (#3319)
This is mucking with dependency so consider this more of a 'Hey, what does everyone think of this'.  Take a look, and this could go back to the drawing board.

This adds some entries to transform_settings to handle declaring these rooms as a hash, for example:
```
safe_room:
  Crossing: XXXX
  Shard: YYYY
lockpick_room_id:
  Crossing: XXXX
  Shard: YYYY
engineering_room:
  Crossing: XXXX
  Shard: YYYY
outfitting_room:
  Crossing: XXXX
  Shard: YYYY
```
This shouldn't impact any existing yamls and should be opaque to any scripts.  I've been running with this for a couple weeks now, swapping between Shard and Crossing, and it has saved me time.

Anyone have any thoughts about this, or ideas on improvement?  Think this'll cause trouble?