Add performance_pause option to make the pause configurable
performance.lic has an fput which waits for rt to be over in it's before_dying, and lich considers a script in the before_dying phase as already stopped so if you start another script that also starts performance it will hang. I haven't figured out a better way around this but making this specific pause configurable will at least provide a workaround for people hitting it.