Allow defining starlight_threshold for trader spells
Second phase of implementation of trader starlight aura handling.

A user sets aura_frequency: to a time period in seconds.  At startup and at each interval afterwards SpellProcess runs DRCA.perc_aura and records the information.  buff_spells and offensive_spells only cast spells marked with a spell_threshold: level if that level is equal to or above the defined level.

If aura_frequency is not set, then all casting is identical to before -- it casts all spells until you run out of starlight, at which point it gets rid of them.