Return to room upon buying, and dampen pin-tapping
Lots of little things today.

Two changes.  One is changing ;sew to return to the room you started in if you have to leave to go buy pins or thread.  Otherwise you continue crafting inside the crafting society.

Second is moving check_pins into the pins section so you don't tap your pins in between every sewing command.  When it gets the pins message, it'll check to see if you've got a pin collection, buy some if you don't, then come back.

I've been running this for about a month and had it run through pins and thread a number of times in that period.