Fix tinker crafting-spell-training and potential fix
First, fixing my obvious if/unless mistake in tinker spell logic on crafting_training_spells.empty? caused by a persistent failure with my brain.

Second, I'm changing the structure of magic_routine a bit, and I'd like to see if we want to do it to the others.

So, there is a case when magic_routine enters with an already prepped training spell and nearly maxed magic mindstates.  It gets past needs_training, casts successfully, but then prepares a new spell even though your mind states are now too high to cast it the next time this is called.  The next time magic_routine would be called it would return on needs_training being nil.

Normally this just means you let the prepped spell time out, but if you have some spells with symbiosis and some without then this could cause you to cast the wrong spell with symbioisis still up.  Rare, but it did occasionally give me a bit of nerve damage when I was still doing that.

Instead, this will store the spell being prepped in the preparing variable rather than making it true/false, then check to see if it needs to be cast before doing a needs_training check.