Reset only modifies a skills baseline (#3400)
Edge cases occur after doing a reset when the list of all skills
is reset; Any script which checks a skill rank or exp level shortly
after a reset will incorrectly read a 0. Instead this sets the baseline
to the current skill level, which is what we're after when doing a reset.